backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
pymongo[srv]>=4.6.1
dnspython>=2.6.1
motor>=3.3.2
pydantic>=2.10.0
pydantic-settings>=2.6.0
torch>=2.6.0
torchvision>=0.21.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
matplotlib>=3.8.0
fpdf2>=2.7.8
python-dotenv>=1.0.1
